+++ src/app/repository/team/interface.tsx
@@ -11,4 +11,6 @@ import {Team} from "../../services/sdk/messages/core.types_pb";
 
 export interface ITeam extends Team.AsObject{
     unread_counter?: number;
+    notify?: boolean;
+    count_unread?: boolean;
 }