+++ wasm/tiny-build.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-GOOS=js GOARCH=wasm tinygo build -ldflags="-s -w" -o river.wasm -target wasm flat/main.go
+GOOS=js GOARCH=wasm tinygo build -o river.wasm -target wasm flat/main.go